+++ src/compressed_tensors/quantization/__init__.py
@@ -17,5 +17,6 @@
 
 from .quant_args import *
 from .quant_config import *
+from .quant_names import *
 from .quant_scheme import *
 from .lifecycle import *
